fix: Added controllers to user and auth
