+++ app/routes/auth.routes.js
@@ -5,66 +5,21 @@ const { validationResult } = require("express-validator");
 const bcrypt = require("bcrypt");
 const auth = require("../middleWare/auth");
 
-const {
-  loginValidation,
-  saltGenerator,
-  passwordGenerator,
-} = require("../utils/utils");
-
+const { loginValidation } = require("../utils/utils");
+const UsersController = require("../controllers/users.controller");
+const authController = require("../controllers/auth.controller");
 /**
  * @route   GET api/auth
- * @desc    Test route
- * @access  Public
+ * @desc    Get user details by token
+ * @access  Private
  */
-router.get("/", auth, async (req, res) => {
-  try {
-    const user = await User.findById(req.user);
-    res.json(user);
-  } catch (err) {
-    console.error(err.message);
-    res.status(500).send("Server Error");
-  }
-});
+router.get("/", auth, authController.auth);
 
 /**
- * @route   POST api/auth
+ * @route   POST api/users
  * @desc    Authenticate User and get token
  * @access  Public
  */
-router.post("/", loginValidation(), async (req, res) => {
-  const errors = validationResult(req);
-
-  if (!errors.isEmpty()) {
-    return res
-      .status(400)
-      .json({ status: false, full_messages: errors.array() });
-  }
-
-  const { email, password } = req.body;
-
-  try {
-    const userLogin = await User.findOne({ email });
-
-    if (!userLogin) {
-      return res
-        .status(400)
-        .json({ status: false, full_messages: "Invalid Credentials" });
-    }
-
-    const isMatch = await bcrypt.compare(password, userLogin.passwordHash);
-
-    if (!isMatch) {
-      return res
-        .status(400)
-        .json({ status: false, full_messages: "Invalid Credentials" });
-    }
-    const user = new User();
-    const token = await user.generateAuthToken();
-    res.status(201).json({ token });
-  } catch (err) {
-    console.error(err.message);
-    res.status(500).send("Server error");
-  }
-});
+router.post("/", loginValidation(), authController.login);
 
 module.exports = router;
